Updated a lot of stuff
