+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   
   <modelVersion>4.0.0</modelVersion>
-  <groupId>me.mrCookieSlime</groupId>
+  <groupId>io.github.thebusybiscuit</groupId>
   <artifactId>ExoticGarden</artifactId>
   <version>UNOFFICIAL</version>
   
@@ -77,7 +77,7 @@
   	<dependency>
   		<groupId>com.github.TheBusyBiscuit</groupId>
   		<artifactId>Slimefun4</artifactId>
-  		<version>83685533f9</version>
+  		<version>e3d84de7b9</version>
   		<scope>provided</scope>
   		<exclusions>
   			<exclusion>
