+++ src/main/java/io/github/thebusybiscuit/exoticgarden/PlantType.java
@@ -1,0 +1,10 @@
+package io.github.thebusybiscuit.exoticgarden;
+
+public enum PlantType {
+
+	BUSH,
+	FRUIT,
+	DOUBLE_PLANT,
+	ORE_PLANT;
+
+}
