--- src/main/java/me/mrCookieSlime/ExoticGarden/PlantType.java
@@ -1,10 +1,0 @@
-package me.mrCookieSlime.ExoticGarden;
-
-public enum PlantType {
-
-	BUSH,
-	FRUIT,
-	DOUBLE_PLANT,
-	ORE_PLANT;
-
-}
