+++ src/main/resources/plugin.yml
@@ -1,11 +1,9 @@
 name: ExoticGarden
 version: ${project.version}
-authors: [mrCookieSlime, Goldentoenail, Joapple, SoSeDiK]
+author: The Slimefun 4 Community
 description: A Slimefun Addon that adds new Plants and dishes to the game
-website: http://TheBusyBiscuit.github.io/
 api-version: 1.13
 
-main: me.mrCookieSlime.ExoticGarden.ExoticGarden
+main: io.github.thebusybiscuit.exoticgarden.ExoticGarden
 
 depend: [Slimefun]
-softdepend: [CS-CoreLib]
